+++ commands_docs/commands_mutable_x.php
@@ -186,5 +186,68 @@ return array(
 
 			"<pre>update table set expire = [expiration] where key >= [key] and key < [range_end] limit 65'536</pre>"
 	),
+
+
+
+	new Cmd(
+			"XNEXPIREAT",
+
+			"XNEXPIREAT key timestamp prefix",
+
+			"EXPIREAT up to 65'536 key-value pairs after <i>key</i>.<br />" .
+			"EXPIREAT ONLY valid pairs, and only if they are matching the <i>prefix</i>.",
+
+			"string",
+			"Last key, if there is second page.",
+
+			"1.3.7.5",
+			"READ + [number of keys] * WRITE",
+			false,
+			true,
+
+			"mutable_x",
+
+			"<pre>set u:001:name  John<br />" .
+			"set u:001:city  LA<br />" .
+			"set u:001:state CA<br />" .
+			"set u:001:phone 1.800.12345678<br />" .
+			"<br />" .
+			"xnget      u:001:       1000 u:001:<br />" .
+			"xnexpireat u:001: 1830290400 u:001:</pre>",
+
+			"<pre>update table set expire_at = [expiration] where key >= [key] and key like '[key]%' limit 65'536</pre>"
+	),
+
+
+
+	new Cmd(
+			"XREXPIREAT",
+
+			"XREXPIREAT key timestamp prefix",
+
+			"EXPIREAT up to 65'536 key-value pairs after <i>key</i>.<br />" .
+			"EXPIREAT ONLY valid pairs, and only if they are less than or equal the <i>range_end</i>.",
+
+			"string",
+			"Last key, if there is second page.",
+
+			"1.3.7.5",
+			"READ + [number of keys] * WRITE",
+			false,
+			true,
+
+			"mutable_x",
+
+			"<pre>set price:2010-01 5<br />" .
+			"set price:2010-02 6<br />" .
+			"set price:2010-03 10<br />" .
+			"set price:2010-04 16<br />" .
+			"set price:2010-05 22<br />" .
+			"<br />" .
+			"xrget      price:2010-02       1000 price:2010-04<br />" .
+			"xrexpireat price:2010-02 1830290400 price:2010-04</pre>",
+
+			"<pre>update table set expire_at = [expiration] where key >= [key] and key < [range_end] limit 65'536</pre>"
+	),
 );